+++ R/gam-support.R
@@ -118,7 +118,9 @@ emm_basis.Gam = function(object, trms, xlev, grid, nboot = 800, ...) {
 ### Many thanks to Maarten Jung for help on sorting-out fixed and random effects
 recover_data.gam = function(object, ...) {
     if (length(object$smooth) > 0) { # get rid of random terms
-        fixnm = sapply(object$smooth, function(s) {ifelse(inherits(s, "random.effect"), NA, s$term)})
+        fixnm = unlist(lapply(object$smooth, function(s) {
+                              if(inherits(s, "random.effect")) NA    else s$term
+            }))
         fixnm = union(.all.vars(delete.response(object$pterms)), fixnm[!is.na(fixnm)])
         object$terms = terms(.reformulate(fixnm, env = environment(terms(object))))
     }